# config/documentation/swagger-admin.json
{
	"swagger": "2.0",
	"info": {
		"title": "HonorVet - Admin API",
		"description": "HonorVet - Admin API Documentation",
		"version": "1.0"
	},
	"produces": ["application/json"],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		"/admin/v1/admin/": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-put",
				"tags": ["Sub Admin"],
				"description": "Update Admin Details",
				"security": [
					{
						"Bearer": []
					}
				],
				"content": "multipart/form-data",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"id": {
									"description": "id GUID",
									"required": false,
									"type": "string",
									"value": ""
								},
								"email": {
									"description": "email",
									"required": false,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": false,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": false
								},
								"accessPermissions": {
									"type": "array",
									"required": false,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-post",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Sub Admin"],
				"description": "Add SuperAdmins or SubAdmins",
				"consumes": "application/json",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"email": {
									"description": "email is required",
									"required": true,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": true,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": true
								},
								"accessPermissions": {
									"type": "array",
									"required": true,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-delete",
				"tags": ["Sub Admin"],
				"description": "Delete Admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/admin/block": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-block-put",
				"tags": ["Sub Admin"],
				"description": "Block Admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"enum": [0, 1],
						"type": "integer",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/recruiter/subRecruiterListing": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-compant-listing-get",
				"tags": ["Recruiter Management"],
				"description": "Get Company users listing",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/subRecruiter/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-sub-recruiter-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get sub-Recruiter users details by id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/companyJobs": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-compant-listing-by-Id-get",
				"tags": ["Recruiter Management"],
				"description": "Get Company users listing By Id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/subRecruiterJobs": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-sub-Recruiter-Jobs-listing-by-Id-get",
				"tags": ["Recruiter Management"],
				"description": "Get subRecruiterJobs users listing By Id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/subRecruiterJobs/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-sub-recruiter-jobs-detail-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get sub-Recruiter users details by id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/candidateHiredRejected": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-sub-Recruiter-Candidate-hired-rejected-listing-by-Id-get",
				"tags": ["Recruiter Management"],
				"description": "Get subRecruiterJobs users listing By Id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "status",
						"in": "query",
						"description": "// 0 for pending 1 for shortlisted 2 for rejected and 3 for on hold",
						"required": false,
						"type": "number",
						"enum": [0, 1, 2,3],
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/candidateHiredRejected/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-candidate-hired-rejected-detail-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get sub-Recruiter users details by id",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/subscriptionHistory/detail": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-candidate-subscriptionHistory-detail-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get Subscription plan history of company",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "Company ID send",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/listing": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-listing-get",
				"tags": ["Recruiter Management"],
				"description": "Get Recruiter listing by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "isAdminApproved",
						"in": "query",
						"description": "0 is pending 1 is approved -1 rejected",
						"required": false,
						"type": "string",
						"enum": [-1, 0, 1],
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/delete": {
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-{id}-delete-delete",
				"tags": ["Recruiter Management"],
				"description": "Delete Recruiter by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/block": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-{id}-block",
				"tags": ["Recruiter Management"],
				"description": "block Recruiter by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "isBlocked",
						"enum": [0, 1],
						"description": "Blocked admin Recruiter",
						"in": "formData",
						"type": "integer",
						"required": false
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/permission": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-{id}-permission-change",
				"tags": ["Recruiter Management"],
				"description": "permission Recruiter by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "isAdminApproved",
						"enum": [0, 1, -1],
						"description": "permission Recruiter by admin",
						"in": "formData",
						"type": "integer",
						"required": true
					},
					{
						"name": "reason",
						"in": "formData",
						"description": "reason",
						"type": "string",
						"value": "",
						"required": false
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/recruiter/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-detail-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get User details by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
	},
	"definitions": {
		"accessPermissions": {
			"type": "object",
			"properties": {
				"module": {
					"type": "string"
				},
				"permission": {
					"enum": [1, 0],
					"description": "0 or 1 to set permission",
					"type": "integer"
				}
			},
			"example": {
				"module": "dashboard",
				"permission": 1
			}
		}
	}
}